+++ RELEASE_NOTES.md
@@ -1,5 +1,8 @@
+#### 2.27.13 - 24.11.2015
+* Using new NuGet v3 protocol to retrieve unlisted packages - https://github.com/fsprojects/Paket/issues/1254
+
 #### 2.27.12 - 24.11.2015
-* WORKAROUND: Working around NuGt v3 protocol issues with unlisted packages - https://github.com/fsprojects/Paket/issues/1254
+* WORKAROUND: Working around NuGet v3 protocol issues with unlisted packages - https://github.com/fsprojects/Paket/issues/1254
 
 #### 2.27.11 - 24.11.2015
 * BUGFIX: Specifying prereleases did not work with pessimistic version constraint - https://github.com/fsprojects/Paket/issues/1252